+++ src/sections/index.js
@@ -1,5 +1,5 @@
 import {default as Cover} from './CoverSection';
-import {default as New} from './NewSection';
+import {default as AboutHacks} from './AboutHacks';
 import {default as About} from './AboutSection';
 import {default as Prize} from './PrizeSection';
 import {default as Schedule} from './ScheduleSection';
@@ -10,4 +10,4 @@ import {default as Contact} from './ContactSection';
 import {default as Team} from './TeamSection';
 
 
-export {Cover,New,About, Prize, Schedule, Theme, Sponsor, Faq, Team, Contact}
+export {Cover,AboutHacks,About, Prize, Schedule, Theme, Sponsor, Faq, Team, Contact}
